module/pyproject.toml
[project]
name = "pdf_to_json"
version = "0.1.0"
description = "Translate scanned magazine PDFs to schema-conformant JSON via local Ollama models."
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "pymupdf>=1.24",
    "Pillow>=10.0",
    "httpx>=0.27",
    "numpy>=1.26",
    "scipy>=1.11",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pdf_to_json"]
